Analysis of microarray experiment
(1) the experiment contains totally eight samples - clumps of human
cells in this case
(2) four samples have been treated with a chemical, and four are
non-treated controls
(3) there are approximately 22000 variables measured on each sample.

What we want to find out is which, if any, of all these 22000 variables
are changing (in medscience we often say ”regulated”) significantly as a
result of the treatment.